### 2022-07-27 #551 @grwhumphries
Submission: stochLAB
Submitting Author Name: Grant Humphries
Submitting Author Github Handle: @grwhumphries
Repository:  https://github.com/HiDef-Aerial-Surveying/stochLAB
Version submitted: 1.1.1
Submission type: Stats
Badge grade: gold
Version accepted: TBD
---

```
Package: stochLAB
Title: Stochastic Collision Risk Model
Version: 1.1.0
    person("Grant", "Humphries", , "grant.humphries@hidefsurveying.co.uk", role = c("aut", "cre")),
    person("Bruno", "Caneco", , "bruno@dmpstats.com", role = "aut"),
    person("Marine Scotland", role = c("fnd", "cph")),
    person("HiDef", role = "cph"),
    person("DMPstats", role = "cph")
  )
Description: Tool for running Collision Risk Models for seabirds on
    offshore wind farms.
License: GPL (>= 3) + file LICENSE
    cli,
    foreach,
    glue,
    logr,
    msm,
    pracma,
    tidyr
    covr
Language: en-GB
LazyDataCompression: bzip2
    https://hidef-aerial-surveying.github.io/stochLAB/
```

- [X] A pre-submission inquiry has been approved in [issue#506](506)

This package is designed for consultants or scientists working in the offshore windfarm industry who are looking to quantify the risks of collision to seabirds

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

Core calculations follow the work developed by [Masden (2015)](https://data.marine.gov.scot/dataset/developing-avian-collision-risk-model-incorporate-variability-and-uncertainty)

    Please include hyperlinked references to all other relevant software.

- NA

Gold

Compliance with a good number of standards beyond those identified as minimally necessary. This will require reviewers and authors to agree on identification of both a minimal subset of necessary standards, and a full set of potentially applicable standards. This aspect may be considered fulfilled if at least one quarter of the additional potentially applicable standards have been met, and should definitely be considered fulfilled if more than one half have been met.
