### 2022-08-04 #551 @mpadge
Thanks for the submission @grwhumphries, and for bearing with us while we use submissions to refine our system. In your case, the above "srr" section reveals a bug which we didn't anticipate. I can now appreciate that our [*Guide for Authors*](https://stats-devguide.ropensci.org/pkgdev.html) of statistical packages isn't clear enough, and needs a sentence at the outset along the lines of:

> All packages submitted for software peer review must document compliance with both General Standards, and at least one set of our category-specific standards.

Your package only documents compliance with the general standards, and needs to include statements of compliance with at least one specific category of standards. Updates to our checking system in response to your submission would now generate the following message:

---

:heavy_multiplication_x: Error: Package documents compliance only with general standards. Statistical packages must document compliance with at least one set of category-specific standards as well.

---

I'll leave you to discuss with @emilyriederer whether you might need to put a "Holding" label on this submission while you address the missing standards. Please accept our apologies for any confusion.
