### 2022-08-08 #551 @emilyriederer
Thanks, @mpadge

Hi @grwhumphries - per Mark's comment, please let us know what sort of timeline you prefer for addressing the standards. I can put a `holding` label in this issue in the meantime if you don't expect it to be a short-term priority. Whatever works best for you!
