### 2022-08-11 e6b3f24
removed unquoted calls to columns in dplyr functions and added globalVariables call to globals.R function. Also tidied up the Rbuildignore file to include non-standard directories used on Github
