+++ R/globals.r
@@ -1,1 +1,12 @@
-utils::globalVariables(c("chord_prof_5MW", "referenceProbs"))
+#utils::globalVariables(c("chord_prof_5MW", "referenceProbs"))
+utils::globalVariables(c("chord_prof_5MW",
+                         "referenceProbs",
+                         "Month",
+                         "month",
+                         "BladeWidth",
+                         "Pitch",
+                         "RotorRadius",
+                         "RotorSpeed",".","!<-"
+                         ))
+
+utils::suppressForeignCheck(c(".", "!<-"))
