+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   Running_stochLAB: Running_stochLAB.html
-last_built: 2022-07-27T20:22Z
+last_built: 2022-08-11T20:38Z
 urls:
   reference: https://hidef-aerial-surveying.github.io/stochLAB/reference
   article: https://hidef-aerial-surveying.github.io/stochLAB/articles
