### 2022-08-12 fb1ac0e
update Rbuildignore to ignore LICENSE file and update DESCRIPTION
