### 2022-08-12 d5d43af
update Rbuildignore to ignore LICENSE file and update DESCRIPTION
