### 2022-08-26 #551 @grwhumphries
@mpadge and @emilyriederer  okay the latest version of `stochLAB` (v1.1.1) has just been pushed and I think that covers out the standards for EDA and PD.
