### 2022-08-29 #551 @mpadge
@emilyriederer One of the statistical categories this package complies with is "*Probability Distributions*", which is currently in [draft-mode only](https://github.com/ropensci/statistical-software-review-book/pull/67). I've implemented a corresponding "dev" branch of [the `srr` package](https://github.com/ropensci-review-tools/srr/pull/32) which the above checks use. These dev branches should be merged by mid-September 2022. Until then please note the following:

1. Automated checks will fail, because there is no way to trigger "dev" branch checks here, but we'll always see that fail and be able to ping the dev checks ourselves to dump results here.
2. Anybody wanting to check the package locally can easily do so by first running:
    ``` r
    remotes::install_github("ropensci-review-tools/srr@dev")
    ```
    All checks will then be re-directed to current "dev" branch of standards including the new categories;
3. We'll strive to get these new standards incorporated **before reviews of this package commence**. If we don't notify here, both @grwhumphries and reviewers will need to be warned that the *Probability Distributions Standards* may change during the review process, and so reviews *may* require iteration. (We will of course strive to avoid this!).

Thanks both, and great to have submissions really pushing our system along! :rocket: :bullettrain_side:
