### 2022-08-29 1248fcf
document returns for functions
