### 2022-08-29 #551 @grwhumphries
@emilyriederer and @mpadge I've just re-run the package check with the srr dev branch and that seems to have passed. I also fixed the issue above (no returns documented) - which I guess is a new check as that wasn't picked up in previous checks?  It should be ready to go again now
