+++ DESCRIPTION
@@ -18,8 +18,7 @@ Description: Tool for running Collision Risk Models for seabirds on
     Ornithology. 
 License: GPL (>= 3)
 Depends: 
-    R (>= 4.0),
-    tidyverse
+    R (>= 4.0)
 Imports: 
     cli,
     dplyr,
