+++ NAMESPACE
@@ -40,3 +40,4 @@ importFrom(stats,quantile)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(stats,sd)
+importFrom(tibble,as_tibble)
