### 2022-10-03 #551 @stephaniehicks
Hi @grwhumphries, great to meet you too. I apologize for my slow response on this. I'm working on identifying reviewers.

@mpadge can you confirm that I have access to `@ropensci-review-bot seeking reviewers`. At the moment, it looks like I don't have access to it and will need you or another editor's help here to initiate the bot. Thanks!
