### 2022-10-04 #551 @mpadge
@stephaniehicks Sorry about that, an oversight on our part. You should now be able to access all bot commands.
