### 2022-10-11 #551 @stephaniehicks
@mpadge just a small note about the stats dev guide:

The Handling Editor Checklist (https://stats-devguide.ropensci.org/pkgsubmission.html#handling-editor-checklist) says to use `@ropensci-review-bot add <@GITHUB_USERNAME> to reviewers` to add a reviewer, but following `@ropensci-review-bot help`, I noticed it is `@ropensci-review-bot assign xxxxx as reviewer` (i.e. need to replace `add` with `assign`).
