### 2022-10-11 #551 @stephaniehicks
Hi everyone, I'd like to introduce Kelly Street (@kstreet13) (https://kellystreet.org) who is an Assistant Professor at UCS. He has kindly agreed to be a reviewer here. However, I wanted to mention that he will need some time to review the guide to go through the content there first. Just a heads up! Thank you again @kstreet13 and let me know if you have any questions!
