### 2022-10-14 #551 @mpadge
Sorry for confusion @stephaniehicks! The bot should recognise any one of "add to reviewers" "add as reviewer", "assign as reviewer", or "assign to reviewers". I think the [preceding glitch](https://github.com/ropensci/software-review/issues/551#issuecomment-1274757490) was because of missing "to reviewers". Happy that all now seems good :smile:
