### 2022-11-05 #551 @kstreet13
Sorry again for the delay! As I said, this was my first time reviewing for ROpenSci and I was a little unclear on the expectations.

---

---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a silver badge
- [ ] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software:
 - **EA2.4** and **EA2.5** concern the use of multitabular input, which is relevant to both `stoch_crm` and `band_crm`.
 - **G5.2**, **G5.2a**, and **G5.2b**. The authors are correct that inputs are checked by the `validate_inputs` scripts, but this is not explicitly tested, as far as I can tell (ie. by having unit tests that give invalid inupts and checking for the correct error message).
 - **G1.5** is not directly applicable, but the authors do claim in the documentation that their implementation of the extended model is "over 30x faster than Masden's code", which is not demonstrated.
 - **PD3.5** seems to concern integration generally, which is a part of the model.

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.
 - **G1.3** Details of the models including statistical terminology are not always explained, but users are directed to appropriate sources, such as Masden (2015) and Band (2012).
 - **G1.4a** Internal functions are not all documented (esp. in `validate_inputs`)

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways:
Function documentation is quite extensive, which may be consequence of how complicated the core methods are, but is nonetheless impressive.

---

### General Review

*The first sentence lays out what the package does, but I'm a little unclear on who the target audience is. I'm guessing it's some combination of people who own wind farms and ornithologists?*
- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case

#### Algorithms

- How well are algorithms encoded?
*The overall structure looks good to me, though some of the details (both of the method and the implementation) are lost on me. I'm guessing the big `for` loop that runs the model a bunch of times with different parameters is the most time-consuming part, so one potential future improvement would be to run those in parallel rather than sequentially.*

- Is the choice of computer language appropriate for that algorithm, and/or envisioned use of package?
*Definitely. This is a stats-heavy model, so R seems like a natural choice.*

- Are aspects of algorithmic scaling sufficiently documented and tested?
*I'm not sure this question is relevant. It (presumably) scales linearly with the number of iterations, but most functions parameters (that may or may not be inferred from data) rather than directly using any data.*

- Are there any aspects of algorithmic implementation which could be improved?
*The parallelization mentioned above could help in cases where users want to run more iterations or try more models. But overall, the examples run pretty quickly, so it seems to be pretty well optimized (and I'm not qualified to comment on the correctness of the implementation, but that seems to checked by the tests).*

- Regardless of actual coverage of tests, are there any fundamental software operations which are not sufficiently expressed in tests?
*I might be missing something, but I don't see any code coverage statistics. That said, tests seem to focus on assessing the correctness of the implementation. This is very important, but misses some of the more granular stuff like making sure every error/warning message works (as in **G5.8**). `test-validate_inputs.R` has some of this (and then some more, commented out), but given that there are SO MANY inputs to the main functions in this package, it definitely seems like these tests should be more extensive.*

- Is there a need for extended tests, or if extended tests exists, have they been implemented in an appropriate way, and are they appropriately documented?
*I think there need to be more tests, but not an explicit set of "extended tests". The additional checks for appropriate warnings/errors, while numerous, should be relatively quick to run.*

#### Visualisation (where appropriate)

- Do visualisations aid the primary purposes of statistical interpretation of results?
*N/A*

- Are there any aspects of visualisations which could risk statistical misinterpretation?
*N/A*

#### Package Design

- Is the package well designed for its intended purpose?
*Mostly. It enables the usage of a few different models for a common purpose and while there are a lot of exported functions, most users will probably only interact with a few. The main issue is the complexity of those functions.*

- In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
*The functions listed on the main documentation page (`stoch_crm` and `band_crm`) are extrememly complex and I think this makes them difficult to use. `stoch_crm` has 31 arguments, which actually hides the number of parameters in the model, as many of these arguments expect data frames containing multiple parameters. I recognize that this isn't the authors' fault (it's a complex model and they have worked to make it more accessible), but the resulting code is still fairly intimidating. I don't know how familiar users will be with some of these specific arguments, but if it is at all possible to provide more reasonable default values, I think that would help.*

- In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
*I'm not familiar with any other packages that handle similar questions.*

- In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
*I think so.*

- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
*A little bit, yes. All of the `*_pars` arguments expect a one-row data frame with column names `"mean"` and `"sd"`, but I think these could accept numeric vectors, as well. You could require names (as in `c(mean = 0, sd = 1)`), or just make it clear in the documentation of these arguments that the first element is assumed to be the mean and the second is the SD.*

---

---

The guidelines don't seem too concerned about this, but my biggest concern with the package is the vignette. This is usually where I start when I want to use a new package, but it's very sparse and doesn't explain much. I would like for the vignette to provide more background and explanation, like the main package documentation. Also, none of the code is actually executed and some of it doesn't even run (there's a stray "messages" that should have been commented out in the Multiscenario Example. This is also present in the README).

I also want to reiterate that if there's anything the authors can do to reduce the number of arguments, provide more suggested default values, or in any way simplify the user experience, that would be greatly appreciated and make the package more approachable.

Some other minor notes:
 - In the main package documentation, there seems to be some filler text that needs to be replaced ("...fill with migration functions..")
 - There are a lot of checks that look like this:
 ```
 if(missing(bird_dens_opt)){
    bird_dens_opt <- "tnorm"
  }
 ```
 but should probably look like this:
 ```
 bird_dens_opt <- match.arg(bird_dens_opt)
 ```
and `match.arg` will automatically treat the first possible value as the default.
 - In those same checks, `stoch_crm` has two checks for `out_period`, but I'm guessing the second one was supposed to be `out_format`.
 - I don't know if this is intentional, but in the vignette and README, you provide the argument `log_file = paste0(getwd(), "scrm_example.log")`. Since this doesn't include a separator it's creating something with the name `[currentdirectory]scrm_example.log` in the parent of the current directory. If that's not the intended behavior, it might be safer to replace this with `file.path(getwd(), "scrm_example.log")`.
 - There are a few commented out functions at the end of `validate_inputs.R` that could probably be removed
