### 2022-12-06 #551 @clairemas0n
## Package Review for 'stockLAB' by clairemas0n

Big apologies for the delay in my review. I also would like to note that this is my first time reviewing a package and it was quite a challenge. With that in mind, I don't think my review as thorough from the software/coding point of view compared to others, but perhaps my contribution is as a potential end user who would use this package to analyse my data. An emerging threat for the seabird I study is a growing number of offshore wind farm proposals popping up in its marine distribution !!

- [X] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html)

### Compliance with Standards

- [X] This package complies with a sufficient number of standards for a SILVER (?) badge

#### The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software:

All seem fine to be non-applicable. However, I can see how G2.16, G2.4c, G2.4d, G2.4e, & G2.9 would be useful in making sure data is correct format and minimising mistakes/bugs.

#### Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

G1.0 & G1.1 are very well documented. It is easy to understand what work stochLAB builds on and how.
G5.1 datasets for tests are very useful for understanding the utility of the functions and the format for data preparation needed to use the package.

I did not find comprehensive documentation on EA1.0, EA1.1, EA1.2, EA1.3

For packages aiming for silver or gold badges:

- [X] This package extends beyond minimal compliance with standards in the following ways: (please describe)

Documentation of functions is excellent.

---

### General Review

It would be helpful to have another sentence or two here to further contextualise the package.

No mention of non-standard dependencies - assumed to be N/A?

- [X] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
A link to report bugs at github/issues was found in ?stochLAB() in R however couldn't find this on the github.com readme version?

- [X] The documentation is sufficient to enable general use of the package beyond one specific use case
It might be helpful to have more information on how to use the outputs -e.g. example graphs or summary tables. I found it difficult to visualise/understand the stoch_crm() output.

The following sections of this template include questions intended to be used as guides to provide general, descriptive responses. Please remove this, and any subsequent lines that are not relevant or necessary for your final review.

#### Package Design

- Is the package well designed for its intended purpose?
Yes. However for seabird biologists to use this package, I think further documentation of how to use the outputs and perhaps how to access wind farm data would be very helpful.

- In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
Yes. However, as there are many arguments in the function, some more guidance on where to obtain some of these parameters would be helpful.

- In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
Unsure.

- In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
Yes.

- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
Unsure. Perhaps outputs for plotting spatially in R e.g. visualising location of wind farms and movement of seabirds?

---

- [X] Does the package comply with the rOpenSci packaging guide?

Very exciting to see this kind of analysis as at my fingertips with this package. I can honestly see myself being able to use these models for my own analysis.
A few things that would (personally) make this more achievable to me would be if more background was provided on the input data/arguments so I could more easily source and format data to use in the functions. And also, if more guidance or examples were provided on how to interpret and use the outputs.

Thank you for the opportunity. Again, sorry for the delay and I hope my thoughts are helpful.
