### 2022-12-07 cc0c010
update for CRAN release
