+++ R/stoch_crm.r
@@ -71,8 +71,8 @@
 #'   requirements.
 #'
 #'   **NOTE:** [generic_fhd_bootstraps] is a list object with generic FHD
-#'   bootstrap estimates for 25 seabird species from [Johnson et al
-#'   (2014)](https://doi.org/10.1111/1365-2664.12191)
+#'   bootstrap estimates for 25 seabird species from Johnson et al
+#'   (2014) \doi{10.1111/1365-2664.12191}
 #'    (see usage in Example Section below).
 #'
 #' @param site_fhd_boots Required only for model Option 4, a data frame similar
