+++ README.Rmd
@@ -54,11 +54,11 @@ For a more detailed overview type `?stochLAB`, once installed!
 
 ## Installation
 
-<!-- You can install the released version of stochLAB from [CRAN](https://CRAN.R-project.org) with: -->
+You can install the released version of stochLAB from [CRAN](https://CRAN.R-project.org) with:
 
-<!-- ``` r -->
-<!-- install.packages("stochLAB") -->
-<!-- ``` -->
+```r 
+install.packages("stochLAB")
+```
 
 You can install the development version with:
 
