+++ README.md
@@ -1,13 +1,15 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
 [![test-coverage](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml)
-[![pkgdown](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgdown.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgdown.yaml)
-[![R-CMD-check](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/R-CMD-check.yaml)
-[![pkgcheck](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgcheck.yaml)
 
 # stochLAB <img src='man/figures/logo.png' align="right" height="139" />
 
 <!-- badges: start -->
+
+[![R-CMD-check](https://github.com/HiDef-Aerial-Surveying/stochLAB/workflows/R-CMD-check/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions)
+[![Codecov test
+coverage](https://codecov.io/gh/HiDef-Aerial-Surveying/stochLAB/branch/master/graph/badge.svg)](https://app.codecov.io/gh/HiDef-Aerial-Surveying/stochLAB?branch=master)
 <!-- badges: end -->
 
 `{stochLAB}` is a tool to run Collision Risk Models (CRMs) for seabirds
@@ -22,15 +24,17 @@ developed by [Masden
 to incorporate variability and uncertainty in the avian collision risk
 model originally developed by [Band
 (2012)](https://www.bto.org/sites/default/files/u28/downloads/Projects/Final_Report_SOSS02_Band1ModelGuidance.pdf).
-
-The `{stochLAB}` package was compiled and curated by DMP Statistics and 
-HiDef Aerial Surveying. 
+The package is for use by individuals modelling collision risk of
+seabirds at offshore wind farms. The primary functions take input
+information on the morphology, behaviour and densities of seabirds as
+well data pertaining to the proposed wind farm (i.e., turbine
+dimensions, speed and number).
 
 Code developed under `{stochLAB}` substantially re-factored and
 re-structured Masden’s (heavily script-based) implementation into a
 user-friendly, streamlined, well documented and easily distributed tool.
 Furthermore, the package lays down the code infrastructure for easier
-incorporation of new functionality, e.g. extra parameter sampling
+incorporation of new functionality, e.g. extra parameter sampling
 features, model expansions, etc.
 
 In addition, previous code underpinning core calculations for the
@@ -44,10 +48,12 @@ For a more detailed overview type `?stochLAB`, once installed!
 
 ## Installation
 
-<!-- You can install the released version of stochLAB from [CRAN](https://CRAN.R-project.org) with: -->
-<!-- ``` r -->
-<!-- install.packages("stochLAB") -->
-<!-- ``` -->
+You can install the released version of stochLAB from
+[CRAN](https://CRAN.R-project.org) with:
+
+``` r
+install.packages("stochLAB")
+```
 
 You can install the development version with:
 
@@ -358,152 +364,85 @@ outputs
 
 ### Band model example
 
-This is an example usage of `band_crm()`. This is for a single species 
[…]
