+++ README.Rmd
@@ -12,13 +12,16 @@ knitr::opts_chunk$set(
   out.width = "100%"
 )
 ```
-[![test-coverage](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml)
 
 # stochLAB <img src='man/figures/logo.png' align="right" height="139" />
 
 <!-- badges: start -->
-[![R-CMD-check](https://github.com/HiDef-Aerial-Surveying/stochLAB/workflows/R-CMD-check/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions)
-[![Codecov test coverage](https://codecov.io/gh/HiDef-Aerial-Surveying/stochLAB/branch/master/graph/badge.svg)](https://app.codecov.io/gh/HiDef-Aerial-Surveying/stochLAB?branch=master)
+[![test-coverage](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/test-coverage.yaml)
+[![pkgdown](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgdown.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgdown.yaml)
+[![R-CMD-check](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/R-CMD-check.yaml)
+[![pkgcheck](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/HiDef-Aerial-Surveying/stochLAB/actions/workflows/pkgcheck.yaml)
+
+
 <!-- badges: end -->
 
 `{stochLAB}` is a tool to run Collision Risk Models (CRMs) for seabirds on offshore wind farms.
