### 2022-12-21 551e474
addtional updates to DESCRIPTION and documentation for ROpenSci review
