+++ R/validate_inputs.R
@@ -420,6 +420,20 @@ validate_inputs <- function(model_options,
 
 
 # ------------------------------------------------------------------------------
+#' Check the fhd against high point of turbine
+#'
+#' This function checks to make sure the bootstrapped flight height distribution
+#' contains information that falls within the maximum rotor height of the turbines.
+#'
+#' @param fhd data frame. The data frame containing the flight height distribution.
+#' see the band_crm function for details
+#' @param tid_off A numeric. The tidal offset in meters
+#' @param air_gap A numeric. The air gap (mean sea level to lower tip of blade) in meters
+#' @param hub_hght A numeric. The height of the center of the turbines (the hub) in meters
+#' @param rtr_rad A numeric. The rotor radius in meters
+#' @param fn A character. The function being evaluated (mCRM or sCRM)
+#'
+#' @noRd
 check_fhd_vs_maxtip <- function(fhd, tid_off, air_gap = NULL, hub_hght = NULL,
                                 rtr_rad, fn){
   fhd_name <- deparse(substitute(fhd))
@@ -451,6 +465,11 @@ check_fhd_vs_maxtip <- function(fhd, tid_off, air_gap = NULL, hub_hght = NULL,
 
 
 # ------------------------------------------------------------------------------
+#' Validate logical value
+#'
+#' Validates if any value is logical
+#' @param x Any value.
+#' @noRd
 val_logical <- function(x){
   obj_name <- deparse(substitute(x))
 
@@ -461,6 +480,14 @@ val_logical <- function(x){
 
 
 # ------------------------------------------------------------------------------
+#' Validate if a constant falls in a range
+#'
+#' Validation for if a value falls within a range and if it is a whole number
+#' @param x A numeric value. The value to test
+#' @param min A numeric value. The minimum extent of the range
+#' @param max A numeric value. The maximum extent of the range
+#' @param check_whole A boolean. If TRUE, tests if the value is a whole number
+#' @noRd
 val_constant <- function(x, min = -Inf, max = Inf, check_whole = FALSE){
 
   obj_name <- deparse(substitute(x))
@@ -492,6 +519,12 @@ val_constant <- function(x, min = -Inf, max = Inf, check_whole = FALSE){
 
 
 # ------------------------------------------------------------------------------
+#' Validation of model options
+#'
+#' Will validate the selected model options for the sCRM
+#'
+#' @param model_options a vector. Model options as a vector (1,2,3,4)
+#' @noRd
 val_model_opts <- function(model_options){
 
   valid_opts <- c('1', '2', '3', '4')
@@ -532,6 +565,15 @@ val_model_opts <- function(model_options){
 
 
 # ------------------------------------------------------------------------------
+#' Validate data frame
+#'
+#' Validates the structure of a data frame. relies on val_df_columns function
+#' @param df A data frame. The data frame input
+#' @param dt_type A character. One of "dstn_pars", "samples", "qtls", "fhd", or "chord_prof"
+#' @param exp_colnames A vector. The expected column names for the data frame
+#' @param single_row A boolean. Whether or not the data frame is a single row
+#' @noRd
+
 val_pars_df <- function(df,
                         dt_type = "dstn_pars",
                         exp_colnames = c("mean", "sd"),
@@ -633,7 +675,14 @@ val_pars_df <- function(df,
 
 
 # ------------------------------------------------------------------------------
-# Checks presence and validity of expected columns in data frame
+#' Validate data frame columns
+#'
+#' Checks presence and validity of expected columns in data frame
+#' @param df A data frame. The data frame input
[…]
