### 2022-12-21 #551 @grwhumphries
## Response to reviewers
Thank you so much @clairemas0n and @kstreet13 ->  your reviews have been really helpful. You've both put a significant amount of work into looking through this and we greatly appreciate it.

I have responded to specific comments as pasted here - hope that works; let me know if there are issues.

See attached response.

### Response to @clairemas0n

I did not find comprehensive documentation on EA1.0, EA1.1, EA1.2, EA1.3
**GH – I have added more information in the README that should hopefully contextualise these points a little better**

A statement of need clearly stating problems the software is designed to solve and its target audience in README
It would be helpful to have another sentence or two here to further contextualise the package.
**GH – I have added some more information here in the description and in the README**

Installation instructions: for the development version of package and any non-standard dependencies in README
No mention of non-standard dependencies - assumed to be N/A?
**GH – I have added a sentence in the README on the package dependencies**

Community guidelines including contribution guidelines in the README or CONTRIBUTING
A link to report bugs at github/issues was found in ?stochLAB() in R however couldn't find this on the github.com readme version?
**GH – This has now been added**

The documentation is sufficient to enable general use of the package beyond one specific use case
It might be helpful to have more information on how to use the outputs -e.g. example graphs or summary tables. I found it difficult to visualise/understand the stoch_crm() output.
**GH – I have added new sections to the README to reflect these points**

Is the package well designed for its intended purpose?
CM - Yes. However for seabird biologists to use this package, I think further documentation of how to use the outputs and perhaps how to access wind farm data would be very helpful.
**GH – I have added new sections to the README to reflect these points**

In relation to External Design: Do exported functions and the relationships between them enable general usage of the package?
CM - Yes. However, as there are many arguments in the function, some more guidance on where to obtain some of these parameters would be helpful.
**GH – I have added new sections to the README to reflect these points**

In relation to External Design: Do exported functions best serve inter-operability with other packages?
CM - Unsure.
**GH – two other packages interact with stochLAB:  mCRM (hidef-aerial-surveying/mCRM) and sCRM (dmpstats/sCRM) -> both have been designed around this package and respective outputs.**

In relation to Internal Design: Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
CM - Unsure. Perhaps outputs for plotting spatially in R e.g. visualising location of wind farms and movement of seabirds?
**GH – Unfortunately, stochLAB is a-spatial and so there is no scope to add visualisation of the locations of wind farms/seabird movement.**

### Response to @kstreet13

EA2.4 and EA2.5 concern the use of multitabular input, which is relevant to both stoch_crm and band_crm.
**GH – data.frame and/or tibble classes are demanded and checked, so have marked these standards as met**

G5.2, G5.2a, and G5.2b. The authors are correct that inputs are checked by the validate_inputs scripts, but this is not explicitly tested, as far as I can tell (ie. by having unit tests that give invalid inupts and checking for the correct error message).
**GH – This could be something implemented in future iterations; the validate_inputs script is very thorough and should cover most issues. Due to the number of parameters (which can’t be helped), the number of possible errors could be substantial and it would be far too time consuming for every issue to be explicitly tested; however, the validate_inputs script makes sure that all data are input in the correct format.**

G1.5 is not directly applicable, but the authors do claim in the documentation that their implementation of the extended model is "over 30x faster than Masden's code", which is not demonstrated.
**GH – Masden’s code is no longer used for collision risk modelling and we didn’t feel it a useful exercise to benchmark this. We have removed the terminology “30x faster” so that it isn’t quantified. The quantification was done by way of experience with older versions of the code**

G1.4a Internal functions are not all documented (esp. in validate_inputs)
**GH – These functions in validate_inputs are now documented**

A statement of need clearly stating problems the software is designed to solve and its target audience in README
KS - The first sentence lays out what the package does, but I'm a little unclear on who the target audience is. I'm guessing it's some combination of people who own wind farms and ornithologists?
**GH – I have added more information in the README now**

How well are algorithms encoded?
KS - The overall structure looks good to me, though some of the details (both of the method and the implementation) are lost on me. I'm guessing the big for loop that runs the model a bunch of times with different parameters is the most time-consuming part, so one potential future improvement would be to run those in parallel rather than sequentially.
**GH – we discussed parallelization of the stochastic process, but because it runs so quickly, we decided against it.**

Testing
**GH - Our feeling on this was that the validate_inputs function does a thorough job of ensuring the functions work as expected/required. Due to the incredible number of possible ways that error messages could be generated, we felt this was the most appropriate way forward. However, future versions of the package could integrate some of these tests.**

Is the package well designed for its intended purpose?
KS - Mostly. It enables the usage of a few different models for a common purpose and while there are a lot of exported functions, most users will probably only interact with a few. The main issue is the complexity of those functions.
**GH – Unfortunately, the complexity is a function of the number of parameters associated with the models; we couldn’t find a way around this.**

In relation to External Design: Do exported functions and the relationships between them enable general usage of the package?
KS - The functions listed on the main documentation page (stoch_crm and band_crm) are extrememly complex and I think this makes them difficult to use. stoch_crm has 31 arguments, which actually hides the number of parameters in the model, as many of these arguments expect data frames containing multiple parameters. I recognize that this isn't the authors' fault (it's a complex model and they have worked to make it more accessible), but the resulting code is still fairly intimidating. I don't know how familiar users will be with some of these specific arguments, but if it is at all possible to provide more reasonable default values, I think that would help.
**GH – agreed that the function is complicated. The package itself is wrapped up inside of two other R Shiny applications, and we have added some more information in the README to help users along with where to get data.**

In relation to Internal Design: Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
KS - A little bit, yes. All of the *_pars arguments expect a one-row data frame with column names "mean" and "sd", but I think these could accept numeric vectors, as well. You could require names (as in c(mean = 0, sd = 1)), or just make it clear in the documentation of these arguments that the first element is assumed to be the mean and the second is the SD.
**GH – Although practically, yes, this could be rearranged so that it accepts the vector argument, the val_pars_df() function which validates this tests if this is a data frame or tibble, and how many rows are in that data frame. It also tests that column names are exactly “mean” or “sd” to ensure consistency.**

The guidelines don't seem too concerned about this, but my biggest concern with the package is the vignette. This is usually where I start when I want to use a new package, but it's very sparse and doesn't explain much. I would like for the vignette to provide more background and explanation, like the main package documentation. Also, none of the code is actually executed and some of it doesn't even run (there's a stray "messages" that should have been commented out in the Multiscenario Example. This is also present in the README).
**GH – The README examples are now executed**

I also want to reiterate that if there's anything the authors can do to reduce the number of arguments, provide more suggested default values, or in any way simplify the user experience, that would be greatly appreciated and make the package more approachable.
**GH – I have added more information in the README now that should hopefully help to clear up some of these issues and make it more accessible. Unfortunately there isn’t a way to limit the number of parameters due to the nature of the model**

In the main package documentation, there seems to be some filler text that needs to be replaced ("...fill with migration functions..")
**GH – These filler texts have been removed**

match.arg functions
**GH – Thanks for catching that, I have integrated that now**

In those same checks, stoch_crm has two checks for out_period, but I'm guessing the second one was supposed to be out_format.
**GH – Thanks for catching that, now changed**

I don't know if this is intentional, but in the vignette and README, you provide the argument log_file = paste0(getwd(), "scrm_example.log"). Since this doesn't include a separator it's creating something with the name [currentdirectory]scrm_example.log in the parent of the current directory. If that's not the intended behavior, it might be safer to replace this with file.path(getwd(), "scrm_example.log").
**GH – Thanks, have updated this in the vignette and README**

There are a few commented out functions at the end of validate_inputs.R that could probably be removed
**GH – Thanks, have removed these**

### Points for clarification
PD3.5 seems to concern integration generally, which is a part of the model.
**GH – Could use some clarification here as I’m not sure how this standard is applicable to the model exactly.**

G1.3 Details of the models including statistical terminology are not always explained, but users are directed to appropriate sources, such as Masden (2015) and Band (2012).
**GH – The main users of the methodology should be familiar with these papers as they are critical for understanding how results need to be discussed for environmental impact assessments. Were there specific terms that needed clarification?**
