### 2022-12-21 7f4cebf
addtional updates to DESCRIPTION and documentation for ROpenSci review
