### 2022-12-21 46dce2e
addtional updates to DESCRIPTION and documentation for ROpenSci review
