### 2022-12-21 e50ebd7
additional updates to DESCRIPTION and documentation for ROpenSci review
