+++ vignettes/Running_stochLAB.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
+```r
 library(stochLAB)
 ```
 
