+++ .github/workflows/pkgdown.yaml
@@ -25,7 +25,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown
+          extra-packages: any::pkgdown, any::withr, local::.
           needs: website
 
       - name: Deploy package
