### 2022-12-21 4bed4d6
Increment version number to 1.1.2
