+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: stochLAB
 Title: Stochastic Collision Risk Model
-Version: 1.1.1
+Version: 1.1.2
 Authors@R: c(
     person("Grant", "Humphries", , "grwhumphries@blackbawks.net", role = c("aut", "cre"),comment = c(ORCID = "0000-0001-5783-9892")),
     person("Bruno", "Caneco", , "bruno@dmpstats.com", role = "aut"),
