+++ man/band_spreadsheet_dt.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{band_spreadsheet_dt}
-\alias{band_spreadsheet_dt}
-\title{Parameter values and outputs from Band's Collision Risk spreadsheet
-("Final_Report_SOSS02_BandSpreadSheetWorkedExampl1.xlsm")}
-\format{
-A list object containing:
-\describe{
-\item{flight_speed}{Bird flight speed, in m/s}
-\item{rotor_radius}{rotor radius, in meters}
-...
-}
-}
-\usage{
-band_spreadsheet_dt
-}
-\description{
-A list object comprising parameter values and outputs from Band's worked example
-for testing the consistency of package functions with the original model and
-its implementation.
-}
-\keyword{datasets}
