+++ man/Day_Length.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/DayLength.r
+\name{Day_Length}
+\alias{Day_Length}
+\title{Total day and night hours per month}
+\usage{
+Day_Length(wf_latitude)
+}
+\arguments{
+\item{wf_latitude}{A decimal value. The latitude of the centroid of the
+windfarm, in degrees.}
+}
+\value{
+data frame with total number of daylight hours and night hours in each
+month at the specified latitude.
+}
+\description{
+Taken from Forsythe et al.(1995) A model comparison for daylength as a
+function of latitude and day of year.  Ecological Modelling. 80: 87 - 95
+}
+\examples{
+    x <- Day_Length(54.6)
+
+}

[…]
