### 2022-12-23 #551 @clairemas0n
Thank you @grwhumphries for taking the time to consider my comments. The README is now very comprehensive and contextualises the package very well. I acknowledge my review has been responded to and I approve acceptance at a gold level. Congratulations @grwhumphries on all your work getting this package to this standard. Very impressive!

> @clairemas0n - would you like to be listed in the DESCRIPTION as a reviewer?
Sorry missed this. I wasn't sure whether my contributions were substantial enough to be considered a reviewer. But happy to be mentioned if they contributed enough to the development of your package. Not worried either way. Thank you @grwhumphries.
