### 2023-02-02 #551 @maurolepore
Dear all,

I'm the EiC on duty and now also acting as handling editor -- thanks @stephaniehicks for your work!

@kstreet13 could you please let me know if you're happy with the author's response and either request further changes or indicate approval with the [reviewer approval template](https://devguide.ropensci.org/approval2template.html)?
