### 2023-02-02 #551 @kstreet13
Sorry! I somehow missed this in December. I think the package is much improved and all of my concerns have been addressed.

Regarding the points for clarification:
 - for PD3.5, I think I was looking at `get_prop_crh_fhd`, which contains the comment `# integration over the rotor's height bands`, but I think the height bands are discrete, so this isn't actually an approximation of a continuous integral.
 - Model terms are sufficiently explained, I think I was just confused by the complexity and not sufficiently familiar with the original papers.

I am happy to accept this package at a *gold* level. I don't think the standards for testing have been met in the way that they are written, but I absolutely agree with @grwhumphries that to do so would be impractical for this package and the current implementation is more than sufficient.

## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 21
