### 2023-02-03 #551 @ropensci-review-bot
Approved! Thanks @grwhumphries for submitting and @kstreet13, @clairemas0n for your reviews! :grin:
