### 2023-02-14 #551 @maelle
@grwhumphries did you meet any difficulty? I'm asking as I'm drafting rOpenSci newsletter I'll send on Friday: if your package is transferred by then, it will be mentioned in this month's newsletter (if not, in next month's so no worries).
