### 2023-02-23 #551 @maelle
@grwhumphries you first need to transfer the repository to rOpenSci via the repository settings (I've just checked that you had accepted the invitation to the ropensci GitHub organization so you have the rights to do that).

This will transfer the repo but lock you out. To get admin access back, you will then write a comment like https://github.com/ropensci/software-review/issues/551#issuecomment-1441504788

Please tell me how it goes + if anything is unclear.
