### 2023-02-23 #551 @grwhumphries
Hi @maelle - I've already transferred the repo over to rOpenSci  (check out ropensci/stochLAB).  Then I wrote `@ropensci-review-bot finalize transfer of ropensci/stochLAB`  - but still getting: `Could not finalize transfer: Could not add owner rights to the ropensci/stochLAB team`
