### 2023-02-24 #551 @maelle
@grwhumphries this is due to a confusion with the "invite me" command, correct? or which instructions do you mean? (I agree the two different syntaxes are confusing)
